LPS-57070 Moves code to API module and export classes to allow other applications extend the search facets
