+++ modules/apps/search/search-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Search Api
+Bundle-SymbolicName: com.liferay.search.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.search.api.*
