+++ modules/apps/search/search-api/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+	compile group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
+}
+
+liferay {
+	deployDir = file("${liferayHome}/osgi/modules")
+}
