+++ modules/apps/search/search-api/build.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<!DOCTYPE project>
+
+<project>
+	<import file="../../../build-module.xml" />
+</project>
