+++ modules/apps/search/search-web/src/META-INF/resources/init.jsp
@@ -89,11 +89,11 @@ page import="com.liferay.portlet.asset.model.AssetRendererFactory" %><%@
 page import="com.liferay.portlet.asset.service.AssetCategoryLocalServiceUtil" %><%@
 page import="com.liferay.portlet.asset.service.AssetEntryLocalServiceUtil" %><%@
 page import="com.liferay.portlet.asset.service.permission.AssetCategoryPermission" %><%@
+page import="com.liferay.search.api.util.SearchFacet" %><%@
+page import="com.liferay.search.api.util.SearchFacetTracker" %><%@
 page import="com.liferay.search.web.constants.SearchPortletKeys" %><%@
 page import="com.liferay.search.web.context.SearchDisplayContext" %><%@
 page import="com.liferay.search.web.facet.AssetEntriesSearchFacet" %><%@
-page import="com.liferay.search.web.util.SearchFacet" %><%@
-page import="com.liferay.search.web.util.SearchFacetTracker" %><%@
 page import="com.liferay.taglib.aui.AUIUtil" %><%@
 page import="com.liferay.taglib.search.ResultRow" %><%@
 page import="com.liferay.taglib.servlet.PipingServletResponse" %><%@
