+++ modules/apps/search/search-web/src/com/liferay/search/web/context/SearchDisplayContext.java
@@ -24,8 +24,8 @@ import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.security.permission.PermissionChecker;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PropsValues;
-import com.liferay.search.web.util.SearchFacet;
-import com.liferay.search.web.util.SearchFacetTracker;
+import com.liferay.search.api.util.SearchFacet;
+import com.liferay.search.api.util.SearchFacetTracker;
 
 import java.util.List;
