+++ modules/apps/search/search-web/src/com/liferay/search/web/facet/FolderSearchFacet.java
@@ -20,7 +20,8 @@ import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.facet.MultiValueFacet;
 import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.search.web.util.SearchFacet;
+import com.liferay.search.api.facet.BaseJSPSearchFacet;
+import com.liferay.search.api.util.SearchFacet;
 
 import javax.portlet.ActionRequest;
