+++ modules/apps/search/search-web/src/com/liferay/search/web/portlet/action/SearchConfigurationAction.java
@@ -20,9 +20,9 @@ import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.ConfigurationAction;
 import com.liferay.portal.kernel.portlet.DefaultConfigurationAction;
 import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.search.api.util.SearchFacet;
+import com.liferay.search.api.util.SearchFacetTracker;
 import com.liferay.search.web.constants.SearchPortletKeys;
-import com.liferay.search.web.util.SearchFacet;
-import com.liferay.search.web.util.SearchFacetTracker;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
